fix datatype

+++ MP2/internal/member/member.go
@@ -196,7 +196,7 @@ func RandomPermutation(arr *[]int64)  {
 	n := len(*arr)
 	for i := 0; i < n; i++ {
 		j := rng.Int63() % int64(i + 1)
-		if i != j {
+		if int64(i) != j {
 			(*arr)[i], (*arr)[j] = (*arr)[j], (*arr)[i] // swap elements
 		}
 	}
